Properly cast to a DWORD for printf.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@15012 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/AppUtils.cpp
@@ -1,6 +1,6 @@
 ﻿// TortoiseSVN - a Windows shell extension for easy version control
 
-// Copyright (C) 2003-2008 - TortoiseSVN
+// Copyright (C) 2003-2009 - TortoiseSVN
 
 // This program is free software; you can redistribute it and/or
 // modify it under the terms of the GNU General Public License
@@ -1000,7 +1000,7 @@ bool CAppUtils::StartShowUnifiedDiff(HWND hWnd, const CTSVNPath& url1, const SVN
 	{
 		sCmd += _T(" /hwnd:");
 		TCHAR buf[30];
-		_stprintf_s(buf, 30, _T("%d"), hWnd);
+		_stprintf_s(buf, 30, _T("%ld"), (DWORD)hWnd);
 		sCmd += buf;
 	}
 
@@ -1036,7 +1036,7 @@ bool CAppUtils::StartShowCompare(HWND hWnd, const CTSVNPath& url1, const SVNRev&
 	{
 		sCmd += _T(" /hwnd:");
 		TCHAR buf[30];
-		_stprintf_s(buf, 30, _T("%d"), hWnd);
+		_stprintf_s(buf, 30, _T("%ld"), (DWORD)hWnd);
 		sCmd += buf;
 	}
